+++ tests/Wpf/Prism.DryIoc.Wpf.Tests/ContainerHelper.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using DryIoc;
 using Prism.DryIoc;
-using Prism.DryIoc.Regions;
 using Prism.Ioc;
 
 namespace Prism.Container.Wpf.Tests
@@ -31,7 +30,5 @@ namespace Prism.Container.Wpf.Tests
         public static Type BaseContainerInterfaceType = typeof(IContainer);
 
         public static Type RegisteredFrameworkException = typeof(ContainerException);
-
-        public static Type RegionNavigationContentLoaderType = typeof(DryIocRegionNavigationContentLoader);
     }
 }
